requirements.txt
sounddevice
scipy
python-dotenv
whisper 
requests
keyboard